Crate nrf_hal_common

Implementation details of the nRF HAL crates. Don’t use this directly, use one of the specific HAL crates instead (nrfXYZ-hal).

Re-exports

pub use crate::clocks::Clocks;

pub use crate::delay::Delay;

pub use crate::rng::Rng;

pub use crate::rtc::Rtc;

pub use crate::timer::Timer;

pub use crate::saadc::Saadc;

pub use crate::spim::Spim;

pub use crate::twim::Twim;

pub use crate::uarte::Uarte;

pub use nrf52840_pac as pac;

Modules

ccm

HAL blocking interface to the AES CCM mode encryption.

clocks

Configuration and control of the High and Low Frequency Clock sources.

comp

HAL interface for the COMP peripheral.

delay

Delays.

ecb

HAL interface to the AES electronic codebook mode encryption.

gpio

gpiote

HAL interface for the GPIOTE peripheral.

i2s

HAL interface for the I2S peripheral.

ieee802154

IEEE 802.15.4 radio

lpcomp

HAL interface for the LPCOMP peripheral.

nvmc

HAL interface to the Non-Volatile Memory Controller (NVMC) peripheral.

ppi

HAL interface for the PPI peripheral.

prelude

pwm

HAL interface to the PWM peripheral.

qdec

HAL interface for the QDEC peripheral.

rng

HAL interface to the RNG peripheral.

rtc

A high level interface for RTC peripherals.

saadc

HAL interface to the SAADC peripheral.

spi

HAL interface to the SPI peripheral.

spim

HAL interface to the SPIM peripheral.

spis

HAL interface to the SPIS peripheral.

target_constants

Length of Nordic EasyDMA differs for MCUs

temp

Temperature sensor interface.

time

Time units.

timer

HAL interface to the TIMER peripheral.

twim

HAL interface to the TWIM peripheral.

twis

HAL interface to the TWIS peripheral.

uarte

HAL interface to the UARTE peripheral.

uicr

HAL interface to the UICR core component.

usbd

wdt

HAL interface to the WDT peripheral.