nrf_hal_common::comp

Struct Comp

pub struct Comp { /* private fields */ }

A safe wrapper around the COMP peripheral.

Implementations

impl Comp

pub fn new<P: CompInputPin>(comp: COMP, input_pin: &P) -> Self

Takes ownership of the COMP peripheral, returning a safe wrapper.

pub fn power_mode(&self, mode: PowerMode) -> &Self

Sets the speed and power mode of the comparator.

pub fn vref(&self, vref: VRef) -> &Self

Sets Vref of the comparator in single ended mode.

pub fn aref_pin<P: CompRefPin>(&self, ref_pin: &P) -> &Self

Sets analog reference pin.

pub fn differential<P: CompRefPin>(&self, ref_pin: &P) -> &Self

Sets comparator mode to differential with external Vref pin.

pub fn hysteresis_threshold_up(&self, value: u8) -> &Self

Upward hysteresis threshold in single ended mode Vup = (value+1)/64*Vref.

pub fn hysteresis_threshold_down(&self, value: u8) -> &Self

Downward hysteresis threshold in single ended mode Vdown = (value+1)/64*Vref.

pub fn hysteresis(&self, enabled: bool) -> &Self

Enables/disables differential comparator hysteresis (50mV).

pub fn enable_interrupt(&self, event: Transition) -> &Self

Enables COMP_LPCOMP interrupt triggering on the specified event.

pub fn disable_interrupt(&self, event: Transition) -> &Self

Disables COMP_LPCOMP interrupt triggering on the specified event.

pub fn enable(&self)

Enables the comparator and waits until it’s ready to use.

pub fn disable(&self)

Disables the comparator.

pub fn is_up(&self) -> bool

Checks if the Up transition event has been triggered.

pub fn is_down(&self) -> bool

Checks if the Down transition event has been triggered.

pub fn is_cross(&self) -> bool

Checks if the Cross transition event has been triggered.

pub fn event_up(&self) -> &EVENTS_UP

Returns reference to Up transition event endpoint for PPI.

pub fn event_down(&self) -> &EVENTS_DOWN

Returns reference to Down transition event endpoint for PPI.

pub fn event_cross(&self) -> &EVENTS_CROSS

Returns reference to Cross transition event endpoint for PPI.

pub fn reset_event(&self, event: Transition)

Marks event as handled.

pub fn reset_events(&self)

Marks all events as handled.

pub fn read(&self) -> CompResult

Returns the output state of the comparator.

pub fn free(self) -> COMP

Consumes self and returns back the raw COMP peripheral.