nrf_hal_common::i2s

Struct I2S

pub struct I2S { /* private fields */ }

Implementations

impl I2S

pub fn new(i2s: I2S_PAC, pins: Pins) -> Self

Takes ownership of the raw I2S peripheral, returning a safe wrapper in controller mode.

pub fn enable(&self) -> &Self

Enables the I2S module.

pub fn disable(&self) -> &Self

Disables the I2S module.

pub fn start(&self) -> &Self

Starts I2S transfer.

pub fn stop(&self) -> &Self

Stops the I2S transfer and waits until it has stopped.

pub fn set_tx_enabled(&self, enabled: bool) -> &Self

Enables/disables I2S transmission (TX).

pub fn set_rx_enabled(&self, enabled: bool) -> &Self

Enables/disables I2S reception (RX).

pub fn set_mck_frequency(&self, freq: MckFreq) -> &Self

Sets MCK generator frequency.

pub fn set_ratio(&self, ratio: Ratio) -> &Self

Sets MCK / LRCK ratio.

pub fn set_sample_width(&self, width: SampleWidth) -> &Self

Sets sample width.

pub fn set_align(&self, align: Align) -> &Self

Sets the sample alignment within a frame.

pub fn set_format(&self, format: Format) -> &Self

Sets the frame format.

pub fn set_channels(&self, channels: Channels) -> &Self

Sets the I2S channel configuration.

pub fn channels(&self) -> Channels

Returns the I2S channel configuration.

pub fn rx<W, B>(self, buffer: B) -> Result<Transfer<B>, Error>
where W: SupportedWordSize, B: WriteBuffer<Word = W> + 'static,

Receives data into the given buffer until it’s filled. Buffer address must be 4 byte aligned and located in RAM. Returns a value that represents the in-progress DMA transfer.

pub fn transfer<W, TxB, RxB>( self, tx_buffer: TxB, rx_buffer: RxB, ) -> Result<TransferFullDuplex<TxB, RxB>, Error>
where W: SupportedWordSize, TxB: ReadBuffer<Word = W> + 'static, RxB: WriteBuffer<Word = W> + 'static,

Full duplex DMA transfer. Transmits the given tx_buffer while simultaneously receiving data into the given rx_buffer until it is filled. The buffers must be of equal size and their addresses must be 4 byte aligned and located in RAM. Returns a value that represents the in-progress DMA transfer.

pub fn tx<W, B>(self, buffer: B) -> Result<Transfer<B>, Error>
where W: SupportedWordSize, B: ReadBuffer<Word = W> + 'static,

Transmits the given tx_buffer. Buffer address must be 4 byte aligned and located in RAM. Returns a value that represents the in-progress DMA transfer.

pub fn set_tx_ptr(&self, addr: u32) -> Result<(), Error>

Sets the transmit buffer RAM start address.

pub unsafe fn set_rx_ptr(&self, addr: u32) -> Result<(), Error>

Sets the receive buffer RAM start address.

pub unsafe fn set_buffersize(&self, n_32bit: u32) -> Result<(), Error>

Sets the size (in 32bit words) of the receive and transmit buffers.

pub fn is_event_triggered(&self, event: I2SEvent) -> bool

Checks if an event has been triggered.

pub fn reset_event(&self, event: I2SEvent)

Marks event as handled.

pub fn enable_interrupt(&self, event: I2SEvent) -> &Self

Enables interrupt triggering on the specified event.

pub fn disable_interrupt(&self, event: I2SEvent) -> &Self

Disables interrupt triggering on the specified event.

pub fn event_stopped(&self) -> &EVENTS_STOPPED

Returns reference to Stopped event endpoint for PPI.

pub fn event_rx_ptr_updated(&self) -> &EVENTS_RXPTRUPD

Returns reference to RxPtrUpdated event endpoint for PPI.

pub fn event_tx_ptr_updated(&self) -> &EVENTS_TXPTRUPD

Returns reference to TxPtrUpdated event endpoint for PPI.

pub fn task_start(&self) -> &TASKS_START

Returns reference to Start task endpoint for PPI.

pub fn task_stop(&self) -> &TASKS_STOP

Returns reference to Stop task endpoint for PPI.

pub fn free(self) -> (I2S_PAC, Pins)

Consumes self and returns back the raw peripheral.