nrf_hal_common::pwm

Struct Pwm

pub struct Pwm<T: Instance> { /* private fields */ }

A safe wrapper around the raw peripheral.

Implementations

impl<T> Pwm<T>
where T: Instance,

pub fn new(pwm: T) -> Pwm<T>

Takes ownership of the peripheral and applies sane defaults.

pub fn set_prescaler(&self, div: Prescaler) -> &Self

Sets the PWM clock prescaler.

pub fn prescaler(&self) -> Prescaler

Sets the PWM clock prescaler.

pub fn set_max_duty(&self, duty: u16) -> &Self

Sets the maximum duty cycle value.

pub fn max_duty(&self) -> u16

Returns the maximum duty cycle value.

pub fn set_period(&self, freq: Hertz) -> &Self

Sets the PWM output frequency.

pub fn period(&self) -> Hertz

Returns the PWM output frequency.

pub fn set_output_pin( &self, channel: Channel, pin: Pin<Output<PushPull>>, ) -> &Self

Sets the associated output pin for the PWM channel.

Modifying the pin configuration while the PWM instance is enabled is not recommended.

pub fn swap_output_pin( &mut self, channel: Channel, pin: Pin<Output<PushPull>>, ) -> Option<Pin<Output<PushPull>>>

Sets the output pin of channel, and returns the old pin (if any).

Modifying the pin configuration while the PWM instance is enabled is not recommended.

pub fn clear_output_pin( &mut self, channel: Channel, ) -> Option<Pin<Output<PushPull>>>

Disables the output pin of channel.

The output pin is returned, if one was previously configured.

Modifying the pin configuration while the PWM instance is enabled is not recommended.

pub fn enable(&self)

Enables the PWM generator.

pub fn disable(&self)

Disables the PWM generator.

pub fn enable_channel(&self, channel: Channel) -> &Self

Enables a PWM channel.

pub fn disable_channel(&self, channel: Channel) -> &Self

Disables a PWM channel.

pub fn enable_group(&self, group: Group) -> &Self

Enables a PWM group.

pub fn disable_group(&self, group: Group) -> &Self

Disables a PWM group.

pub fn set_load_mode(&self, mode: LoadMode) -> &Self

Cofigures how a sequence is read from RAM and is spread to the compare register.

pub fn load_mode(&self) -> LoadMode

Returns how a sequence is read from RAM and is spread to the compare register.

pub fn set_counter_mode(&self, mode: CounterMode) -> &Self

Selects operating mode of the wave counter.

pub fn counter_mode(&self) -> CounterMode

Returns selected operating mode of the wave counter.

pub fn set_step_mode(&self, mode: StepMode) -> &Self

Selects source for advancing the active sequence.

pub fn step_mode(&self) -> StepMode

Returns selected source for advancing the active sequence.

pub fn set_duty_on_common(&self, duty: u16)

Sets duty cycle (15 bit) for all PWM channels. Will replace any ongoing sequence playback.

pub fn set_duty_off_common(&self, duty: u16)

Sets inverted duty cycle (15 bit) for all PWM channels. Will replace any ongoing sequence playback.

pub fn duty_on_common(&self) -> u16

Returns the common duty cycle value for all PWM channels in Common load mode.

pub fn duty_off_common(&self) -> u16

Returns the inverted common duty cycle value for all PWM channels in Common load mode.

pub fn set_duty_on_group(&self, group: Group, duty: u16)

Sets duty cycle (15 bit) for a PWM group. Will replace any ongoing sequence playback.

pub fn set_duty_off_group(&self, group: Group, duty: u16)

Sets inverted duty cycle (15 bit) for a PWM group. Will replace any ongoing sequence playback.

pub fn duty_on_group(&self, group: Group) -> u16

Returns duty cycle value for a PWM group.

pub fn duty_off_group(&self, group: Group) -> u16

Returns inverted duty cycle value for a PWM group.

pub fn set_duty_on(&self, channel: Channel, duty: u16)

Sets duty cycle (15 bit) for a PWM channel. Will replace any ongoing sequence playback and the other channels will return to their previously set value.

pub fn set_duty_off(&self, channel: Channel, duty: u16)

Sets inverted duty cycle (15 bit) for a PWM channel. Will replace any ongoing sequence playback and the other channels will return to their previously set value.

pub fn duty_on(&self, channel: Channel) -> u16

Returns the duty cycle value for a PWM channel.

pub fn duty_off(&self, channel: Channel) -> u16

Returns the inverted duty cycle value for a PWM group.

pub fn set_loop(&self, mode: Loop)

Sets number of playbacks of sequences.

pub fn one_shot(&self) -> &Self

Looping disabled (stop at the end of the sequence).

pub fn loop_inf(&self) -> &Self

Loops playback of sequences indefinitely.

pub fn repeat(&self, times: u16) -> &Self

Sets number of playbacks of sequences.

pub fn set_seq_refresh(&self, seq: Seq, periods: u32) -> &Self

Sets number of additional PWM periods between samples loaded into compare register.

pub fn set_seq_end_delay(&self, seq: Seq, periods: u32) -> &Self

Sets number of additional PWM periods after the sequence ends.

pub fn start_seq(&self, seq: Seq)

Loads the first PWM value on all enabled channels from a sequence and starts playing that sequence. Causes PWM generation to start if not running.

pub fn next_step(&self)

Steps by one value in the current sequence on all enabled channels, if the NextStep step mode is selected. Does not cause PWM generation to start if not running.

pub fn stop(&self)

Stops PWM pulse generation on all channels at the end of current PWM period, and stops sequence playback.

pub fn load<B0, B1>( self, seq0_buffer: Option<B0>, seq1_buffer: Option<B1>, start: bool, ) -> Result<PwmSeq<T, B0, B1>, (Error, Pwm<T>, Option<B0>, Option<B1>)>
where B0: ReadBuffer<Word = u16> + 'static, B1: ReadBuffer<Word = u16> + 'static,

Loads the given sequence buffers and optionally (re-)starts sequence playback. Returns a PemSeq, containing Pwm<T> and the buffers.

pub fn enable_interrupt(&self, event: PwmEvent) -> &Self

Enables interrupt triggering on the specified event.

pub fn disable_interrupt(&self, event: PwmEvent) -> &Self

Disables interrupt triggering on the specified event.

pub fn is_event_triggered(&self, event: PwmEvent) -> bool

Checks if an event has been triggered.

pub fn reset_event(&self, event: PwmEvent)

Marks event as handled.

pub fn event_stopped(&self) -> &EVENTS_STOPPED

Returns reference to Stopped event endpoint for PPI.

pub fn event_loops_done(&self) -> &EVENTS_LOOPSDONE

Returns reference to LoopsDone event endpoint for PPI.

pub fn event_pwm_period_end(&self) -> &EVENTS_PWMPERIODEND

Returns reference to PwmPeriodEnd event endpoint for PPI.

pub fn event_seq0_end(&self) -> &EVENTS_SEQEND

Returns reference to Seq0 End event endpoint for PPI.

pub fn event_seq1_end(&self) -> &EVENTS_SEQEND

Returns reference to Seq1 End event endpoint for PPI.

pub fn event_seq0_started(&self) -> &EVENTS_SEQSTARTED

Returns reference to Seq0 Started event endpoint for PPI.

pub fn event_seq1_started(&self) -> &EVENTS_SEQSTARTED

Returns reference to Seq1 Started event endpoint for PPI.

pub fn task_start_seq0(&self) -> &TASKS_SEQSTART

Returns reference to Seq0 Start task endpoint for PPI.

pub fn task_start_seq1(&self) -> &TASKS_SEQSTART

Returns reference to Seq1 Started task endpoint for PPI.

pub fn task_next_step(&self) -> &TASKS_NEXTSTEP

Returns reference to NextStep task endpoint for PPI.

pub fn task_stop(&self) -> &TASKS_STOP

Returns reference to Stop task endpoint for PPI.

pub fn split_channels( &self, ) -> (PwmChannel<'_, T>, PwmChannel<'_, T>, PwmChannel<'_, T>, PwmChannel<'_, T>)

Returns individual handles to the four PWM channels.

pub fn split_groups(&self) -> (PwmGroup<'_, T>, PwmGroup<'_, T>)

Returns individual handles to the two PWM groups.

pub fn free(self) -> (T, Pins)

Consumes self and returns back the raw peripheral.