nrf_hal_common::timer

Struct Timer

pub struct Timer<T, U = OneShot>(/* private fields */);

Interface to a TIMER instance.

Right now, this is a very basic interface. The timer will always be hardcoded to a frequency of 1 MHz and 32 bits accuracy.

CC[0] is used for the current/most-recent delay period and CC[1] is used to grab the current value of the counter at a given instant.

Implementations

impl<T> Timer<T, OneShot>
where T: Instance,

pub fn one_shot(timer: T) -> Timer<T, OneShot>

pub fn new(timer: T) -> Timer<T, OneShot>

impl<T> Timer<T, Periodic>
where T: Instance,

pub fn periodic(timer: T) -> Timer<T, Periodic>

impl<T, U> Timer<T, U>
where T: Instance,

pub const TICKS_PER_SECOND: u32 = 1_000_000

pub fn into_periodic(self) -> Timer<T, Periodic>

pub fn into_oneshot(self) -> Timer<T, OneShot>

pub fn free(self) -> T

Return the raw interface to the underlying timer peripheral.

pub fn read(&self) -> u32

Return the current value of the counter, by capturing to CC[1].

pub fn reset_event(&self)

Reset the interrupt event flag

pub fn enable_interrupt(&mut self)

Enables the interrupt for this timer.

Enables an interrupt that is fired when the timer reaches the value that is given as an argument to start.

Note that the interrupt also has to be unmasked in the NVIC, or the handler won’t get called.

pub fn disable_interrupt(&mut self)

Disables the interrupt for this timer.

Disables an interrupt that is fired when the timer reaches the value that is given as an argument to start.

Note that the interrupt also has to be unmasked in the NVIC, or the handler won’t get called.

pub fn start(&mut self, cycles: u32)

Starts the timer.

The timer will run for the given number of cycles, then it will stop and reset.

pub fn reset_if_finished(&mut self) -> bool

If the timer has finished, resets it and returns true.

Returns false if the timer is still running.

pub fn delay(&mut self, cycles: u32)

Starts the timer for the given number of cycles and waits for it to finish.

pub fn task_start(&self) -> &TASKS_START

Returns reference to the START task endpoint for PPI. Starts timer.

pub fn task_stop(&self) -> &TASKS_STOP

Returns reference to the STOP task endpoint for PPI. Stops timer.

pub fn task_count(&self) -> &TASKS_COUNT

Returns reference to the COUNT task endpoint for PPI. Increments timer (counter mode only).

pub fn task_clear(&self) -> &TASKS_CLEAR

Returns reference to the CLEAR task endpoint for PPI. Clears timer.

pub fn task_capture_cc0(&self) -> &TASKS_CAPTURE

Returns reference to the CC[0] CAPTURE task endpoint for PPI. Captures timer value to the CC[0] register.

pub fn task_capture_cc1(&self) -> &TASKS_CAPTURE

Returns reference to the CC[1] CAPTURE task endpoint for PPI. Captures timer value to the CC[1] register.

pub fn task_capture_cc2(&self) -> &TASKS_CAPTURE

Returns reference to the CC[2] CAPTURE task endpoint for PPI. Captures timer value to the CC[2] register.

pub fn task_capture_cc3(&self) -> &TASKS_CAPTURE

Returns reference to the CC[3] CAPTURE task endpoint for PPI. Captures timer value to the CC[3] register.

pub fn event_compare_cc0(&self) -> &EVENTS_COMPARE

Returns reference to the CC[0] COMPARE event endpoint for PPI. Generated when the counter is incremented and then matches the value specified in the CC[0] register.

pub fn event_compare_cc1(&self) -> &EVENTS_COMPARE

Returns reference to the CC[1] COMPARE event endpoint for PPI. Generated when the counter is incremented and then matches the value specified in the CC[1] register.

pub fn event_compare_cc2(&self) -> &EVENTS_COMPARE

Returns reference to the CC[2] COMPARE event endpoint for PPI. Generated when the counter is incremented and then matches the value specified in the CC[2] register.

pub fn event_compare_cc3(&self) -> &EVENTS_COMPARE

Returns reference to the CC[3] COMPARE event endpoint for PPI. Generated when the counter is incremented and then matches the value specified in the CC[3] register.